feat: add private bank tag web service

+++ site/test/setup.js
@@ -2,7 +2,7 @@ import { afterEach, beforeEach, vi } from "vitest";
 import { pubsub } from "../src/data/pubsub";
 
 beforeEach(() => {
-  localStorage.clear();
+  if (typeof localStorage !== "undefined") localStorage.clear();
   pubsub.subscribers = new Map();
   pubsub.unpublishAll();
 });
